codex: move executor http streaming into client
